Improve PIM display
Added another text line to the display, because there is available space. 
Dimmed the colors of less important info. Moved less important info to the bottom. 
Put preset and voice numbers on a separate line so there is more room for the preset name on the first line. Also always show the active voice number. 
Highlight the menu item in yellow when scrolling buttons change values. This is so you know to press the Y button if you want to stop scrolling values and scroll the menu instead.